refactor: clean up TravelDuration component by removing unused maxDistance logic and adjust form submission feat: add max distance input handling in VelorouteDetails component
